Refactored boards related fnctions
